Add event Trigger API
idk why but my ide is acting up so there are probably errors that ill fix later. Anyway, I added a static field in Choreo.java which stores the current trajectory that the robot is on. Calling choreoSwerveCommand() will update this to the trajectory it is given, and then, when finished, put it back to an empty trajectory. There is also now a new static event() method in Choreo.java which returns a trigger for if the robot is currently on a trajectory given as an argument.

holy guacamole i forgot a semicolon

imported trigger

i mightve jumped the gun but i was bored so heres another event function. is this good

more methodsgit add .! more

le fixes

Add list of keyboard shortcuts to document settings (#447)

Co-authored-by: Jacob Trentini <dev@jacob2.dev>

Feature: Context Menu (#448)

* add context menu with popover

* format

* silence error with conflicting mouseevent

* format

* works but only if you hot reload (fixing with next commit)

* step 1 out of predicted 2: migrate local state to uiState

* step 2/2: move d3 callback to overlaywaypoint

* format

* format and refactorings

* address @calcmogul's concerns

Upgrade TrajoptLib to use larger solver tolerance (#451)

organization and comments

renaming

cleanup java impl

fix docs

format